fix: import param order

+++ examples/rust/src/hello.rs
@@ -2,7 +2,7 @@
 
 #[link(wasm_import_module = "env")]
 extern "C" {
-    fn print_utf8(location: i32, len: i32);
+    fn print_utf8(location: i64, len: i32);
 }
 
 const ARG: &[u8] = &[0u8; 100];
@@ -23,6 +23,6 @@ pub unsafe extern "C" fn hello(len: i32) {
     let res = format!("Hello, {}!", arg).as_bytes().to_vec();
 
     let len = res.len() as i32;
-    let ptr = res.leak().as_ptr() as i32;
+    let ptr = res.leak().as_ptr() as i64;
     print_utf8(ptr, len);
 }
